+++ pkg/ui/react-app/src/pages/status/Status.tsx
@@ -92,7 +92,6 @@ const Status: FC<RouteComponentProps & PathPrefixProps> = ({ pathPrefix = '' }) 
       {[
         { fetchResult: useFetch<Record<string, string>>(`${path}/status/runtimeinfo`), title: 'Runtime Information' },
         { fetchResult: useFetch<Record<string, string>>(`${path}/status/buildinfo`), title: 'Build Information' },
-        { fetchResult: useFetch<Record<string, string>>(`${path}/alertmanagers`), title: 'Alertmanagers' },
       ].map(({ fetchResult, title }) => {
         const { response, isLoading, error } = fetchResult;
         return (