+++ CHANGELOG.md
@@ -16,7 +16,7 @@ We use *breaking* word for marking changes that are not backward compatible (rel
 * [#2665](https://github.com/thanos-io/thanos/pull/2665) Swift: fix issue with missing Content-Type HTTP headers.
 - [#2800](https://github.com/thanos-io/thanos/pull/2800) Query: Fix handling of `--web.external-prefix` and `--web.route-prefix`
 
-### Changed
+### Added
 - [#2832](https://github.com/thanos-io/thanos/pull/2832) ui: React: Add runtime and build info page
 
 ## [v0.14.0](https://github.com/thanos-io/thanos/releases) - IN PROGRESS